requirements.txt
Django==4.2.2
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow==10.0.0
drf-yasg==1.21.6
djoser==2.2.0
